fix: multiple NumPy alignment bug fixes
Bug fixes addressed in this session:
- BUG-19: np.negative applied abs() then negated (always returned negative)
- BUG-20: np.positive applied abs() instead of identity
- BUG-18: np.convolve NullReferenceException
- BUG-15: np.abs changed int dtype to Double
- BUG-24: np.sign changed int32 dtype to Double
- BUG-25: power(int, float) returned int instead of float64
- BUG-13: np.linspace returned float32 instead of float64
- BUG-29: Mixed-type comparison SIMD loaded wrong memory (disabled SIMD for mixed types)

Additional fixes:
- Dot product Int64 to Int32 cast error
- Transpose dimension handling
- NPTypeCode GetGroup/GetPriority methods added

All fixes verified against NumPy 2.4.2 behavior.

+++ src/NumSharp.Core/Creation/np.linspace.cs
@@ -17,7 +17,7 @@ namespace NumSharp
         /// <param name="num">Number of samples to generate. Default is 50. Must be non-negative.</param>
         /// <param name="endpoint">If True, stop is the last sample. Otherwise, it is not included. Default is True.</param>
         /// <param name="dtype">The type of the output array. If dtype is not given, infer the data type from the other input arguments.</param>
-        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html</remarks>
+        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html
         public static NDArray linspace(double start, double stop, int num, bool endpoint, Type dtype)
         {
             return linspace(start, stop, num, endpoint, (dtype ?? typeof(double)).GetTypeCode());
@@ -33,10 +33,11 @@ namespace NumSharp
         /// <param name="num">Number of samples to generate. Default is 50. Must be non-negative.</param>
         /// <param name="endpoint">If True, stop is the last sample. Otherwise, it is not included. Default is True.</param>
         /// <param name="dtype">The type of the output array. If dtype is not given, infer the data type from the other input arguments.</param>
-        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html</remarks>
+        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html
         public static NDArray linspace(float start, float stop, int num, bool endpoint, Type dtype)
         {
-            return linspace(start, stop, num, endpoint, (dtype ?? typeof(float)).GetTypeCode());
+            // NumPy: linspace always returns float64 by default, regardless of input types
+            return linspace(start, stop, num, endpoint, (dtype ?? typeof(double)).GetTypeCode());
         }
 
         /// <summary>
@@ -49,9 +50,10 @@ namespace NumSharp
         /// <param name="num">Number of samples to generate. Default is 50. Must be non-negative.</param>
         /// <param name="endpoint">If True, stop is the last sample. Otherwise, it is not included. Default is True.</param>
         /// <param name="typeCode">The type of the output array. If dtype is not given, infer the data type from the other input arguments.</param>
-        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html</remarks>
-        public static NDArray linspace(float start, float stop, int num, bool endpoint = true, NPTypeCode typeCode = NPTypeCode.Single)
+        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html
+        public static NDArray linspace(float start, float stop, int num, bool endpoint = true, NPTypeCode typeCode = NPTypeCode.Double)
         {
+            // NumPy: linspace always returns float64 by default, regardless of input types
             return linspace((double)start, (double)stop, num, endpoint, typeCode);
         }
 
@@ -65,7 +67,7 @@ namespace NumSharp
         /// <param name="num">Number of samples to generate. Default is 50. Must be non-negative.</param>
         /// <param name="endpoint">If True, stop is the last sample. Otherwise, it is not included. Default is True.</param>
         /// <param name="typeCode">The type of the output array. If dtype is not given, infer the data type from the other input arguments.</param>
-        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html</remarks>
+        /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.linspace.html
         public static NDArray linspace(double start, double stop, int num, bool endpoint = true, NPTypeCode typeCode = NPTypeCode.Double)
         {
             if (typeCode == NPTypeCode.Empty)
